api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
sqlalchemy==1.4.50
databases==0.8.0
asyncpg==0.28.0
aiosqlite==0.19.0
python-dotenv==1.0.0
httpx==0.25.0
python-dateutil==2.8.2
psutil==5.9.5
email-validator==2.2.0 